fix: #77 work around an apparent mismatch between IA and cdx_toolkit
The cdx_toolkit.warc's fetch_wb_warc(), which is called by
CaptureObject.fetch_warc_record, applies urllib.parse.quote() to the
entire URL. However in the case URLs with multiple "?" in them that
does not match where web.archive.org expects.

For URLs with single "?"  it doesn't appear to make a difference.

This works around that, changing how quote() works, so as to keep the
change as localized as possible.